fix: type signature for ImportDocumentResponse
When given return_doc=true in the query to the bulk import API, Typesense
returns a valid JSON object in the `document` field of the response.

On failure (for example invalid JSON), the raw string from the request is
echoed in the response instead.

This change makes the Go SDK compatible with that behavior.